+++ modules/fluent-forms/class-fluent-forms.php
@@ -130,10 +130,15 @@ class AnylabelWP_Fluent_Forms
             <h2><?php esc_html_e('Fluent Forms Settings', 'anylabelwp-plugin'); ?></h2>
             <table class="form-table">
                 <tr valign="top">
-                    <th scope="row"><?php esc_html_e('Custom Fluent Forms Logo URL', 'anylabelwp-plugin'); ?></th>
+                    <th scope="row"><?php esc_html_e('Custom Fluent Forms Logo', 'anylabelwp-plugin'); ?></th>
                     <td>
-                        <input type="text" name="anylabelwp_fluent_forms_logo_url"
-                            value="<?php echo esc_attr(get_option('anylabelwp_fluent_forms_logo_url')); ?>" />
+                        <?php 
+                        \AnylabelWP\Loader::render_logo_selector(
+                            'anylabelwp_fluent_forms_logo_url',
+                            get_option('anylabelwp_fluent_forms_logo_url'),
+                            'forms'
+                        ); 
+                        ?>
                     </td>
                 </tr>
             </table>